% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_logics.R
\name{apply_logics}
\alias{apply_logics}
\title{Apply case_when logics (e.g. policy, censoring) to clones}
\usage{
apply_logics(clones, logics)
}
\arguments{
\item{clones}{A list of data frame. Each element of the list represents
each treatment arm.}

\item{logics}{A nested list. Each element of outer list represents each
treatment arm. Each element of inner list represents each new variable
to be created by applying logics. Each element of inner list containts a
character vector that represents a sequence of logics to be passed into
\code{case_when()} call to determine a value of the new variable.}
}
\value{
A list of data frame that each data frame include new variables
created by the provided logics.
}
\description{
Apply case_when logics (e.g. policy, censoring) to clones
}
