% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clone_arms.R
\name{clone_arms}
\alias{clone_arms}
\title{Duplicate data frame for each treatment arm to emulate}
\usage{
clone_arms(data, arms)
}
\arguments{
\item{data}{Input data frame that contains all observations of interest.
Each row represents an observation, and columns include
observation identifiers, binary treatment variable (0/1),
time to treatement (continuous), binary outcome variable (0/1),
observed followup time (continuous), and covariates.}

\item{arms}{Character vector that each element represents each arm's name.}
}
\value{
A list of data frame.
Each element of list is associated with each arm.
}
\description{
Duplicate data frame for each treatment arm to emulate
}
\examples{
data(lungcancer)
clones <- clone_arms(lungcancer, c("Control", "Surgery"))
}
