% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonecensorweighting-package.R
\docType{package}
\name{clonecensorweighting-package}
\alias{clonecensorweighting}
\alias{clonecensorweighting-package}
\title{clonecensorweighting: Infrastructure for Clone-Censor-Weighting Analyses}
\description{
Provides a starter package for clone-censor-weighting workflows used in target trial emulation. The package currently includes data ingestion helpers, a minimal data cloning routine, and utilities for constructing survival responses.
}
\author{
\strong{Maintainer}: Sang Ho Park \email{shstat1729@gmail.com}

Authors:
\itemize{
  \item Sang Ho Park \email{shstat1729@gmail.com}
}

}
\keyword{internal}
