% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{create_final_data}
\alias{create_final_data}
\title{Create training data for censoring probability estimation}
\usage{
create_final_data(
  clones,
  clone_followup,
  clone_outcome,
  clone_censoring,
  col_ids,
  timestamp_start = "Tstart",
  id = "ID",
  timestamp_stop = "Tstop"
)
}
\arguments{
\item{clones}{A list of data frame. Each element of the list represents
each treatment arm. This version of clones must contain a column that
represents an emulated follow up time (corresponding to \code{clone_followup}
argument), an emulated outcome (correspodning to \code{clone_outcome}), and a
binary indicator variable that represents whether the observation violates
arm's policy or not (corresponding to \code{clone_censoring} argument).}

\item{clone_followup}{A column name that represents the emulated follow up
time in each arm of clones. The variable should exists in each element
data frame of \code{clones} argument.}

\item{clone_outcome}{A column name that represents the emulated outcome
in each arm of clones. The variable should exists in each element
data frame of \code{clones} argument, and the variable value should be binary
(0 or 1).}

\item{clone_censoring}{A column name that represent whether the observation
violates arm's policy or not. The variable should exists in each element
data frame of \code{clones} argument, and the variable value should be binary
(0 or 1).}

\item{col_ids}{A vector of column names that a combination of their values
uniquely identifies each observation.}

\item{timestamp_start}{A new variable name to denote start time of each
subrecord of observations in a long-form data.}

\item{id}{A new variable name for a unique observation identifier, to
represents that multiple rows in output data frame is associated with the
same observation.}

\item{timestamp_stop}{A new variable name to denote end time of each
subrecord of observations in a long-form data.}
}
\value{
A list of long-form data frames. Each data frame represents each
clone arm. Each row of the long-form data frame represents a subrecord of
each observation associated with each specific time interval. The first
subrecord starts with time 0, and the rows are expanded up to
\code{clone_followup}, where cut times are determined by \code{t_events} argument.
}
\description{
Create training data for censoring probability estimation
}
