% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{create_timestamp_table}
\alias{create_timestamp_table}
\title{Create timestamp table}
\usage{
create_timestamp_table(clones, clone_followup)
}
\arguments{
\item{clones}{A list of data frame. Each element of the list represents
each treatment arm. This version of clones must contain a column that
represents a emulated follow up time.}

\item{clone_followup}{A column name in emulcated clone (i.e. \code{clones})
that represents the emulated follow up time in cloned data frame.}
}
\value{
A data frame with two columns: \code{tevent} and \code{ID_t}.
\code{tevent} represents a timestamp that outcome event can occur based on
observed data. \code{ID_t} represents an enumerated identifier of each
timestamp, from 1 to n where n represents the number of unique \code{tevent}
value.
}
\description{
Create timestamp table
}
