% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{split_at_timestamp}
\alias{split_at_timestamp}
\title{Split each observation into multiple subrecords at each time cut}
\usage{
split_at_timestamp(
  clones,
  clone_followup,
  t_events,
  event,
  timestamp_start = "Tstart",
  id = "ID"
)
}
\arguments{
\item{clones}{A list of data frame. Each element of the list represents
each treatment arm. This version of clones must contain a column that
represents a emulated follow up time (corresponding to \code{clone_followup}
argument) and an event of interest (corresponding to \code{event} argument).}

\item{clone_followup}{A column name in emulcated clone (i.e. \code{clones})
that represents the emulated follow up time in cloned data frame.}

\item{t_events}{A vector of timestamp that outcome event can occur based on
observed data.}

\item{event}{A variable name of an event of interest. The variable should
exists in each data frame that is an element of \code{clones} argument, and
the variable value should be a binary (0 or 1).}

\item{timestamp_start}{A new variable name to denote start time.}

\item{id}{A new variable name for a unique observation identifier, to
represents that multiple rows in output data frame is associated with the
same observation.}
}
\value{
A list of long-form data frames. Each data frame represents each
clone arm. Each row of the long-form data frame represents a subrecord of
each observation associated with each specific time interval. The first
subrecord starts with time 0, and the rows are expanded up to
\code{clone_followup}, where cut times are determined by \code{t_events} argument.
}
\description{
Split each observation into multiple subrecords at each time cut
}
